+++ Component/Renderer/dxrg/D3D12Fence.h
@@ -11,7 +11,7 @@ public:
 	virtual uint64 getCompletedValue() const override;
 	virtual uint64 getCurrentFenceValue() const override;
 	virtual uint64 incrementFenceValue() override;
-	virtual void waitForCompletion(uint32 fence) const override;
+	virtual void waitForCompletion(uint64 fence) const override;
 	virtual ~D3D12Fence() override = default;
 private:
 	WRL::ComPtr<ID3D12Fence> _pFence;
